Update version in appdata.xml
There should probably be some automated process that does this.
Also, the changed names of this appdata & the desktop file to match Flatpak(?) naming requirements was something that made me, and probably a few other packagers need to change things to find these newly named files. Perhaps we could have both the Flatpak and normally named files so that this doesn't cause problems?